+++ wave.h
@@ -47,11 +47,11 @@ class WAVE {
 
     // Get the wave sample specified by index
     uint8_t sample(uint8_t idx);
-    // Get the wave sample specified by index in Q14.2 format
+    // Get the wave sample specified by index in Q8.8 format
     uint8_t sample(uint16_t idx);
 
-    // Compute the wave steps as Q6.2
-    uint8_t getStep(uint16_t freq);
+    // Compute the wave steps as Q8.8
+    uint16_t getStep(uint16_t freq);
 };
 
 #endif /* WAVE_H */
